Add files via upload
Excel's Macro Recorder Tool, using this excel tool using a base table, I implemented headers and currency formatting for Dollars.
Just using the ctrl+j command we can automate the processes in this case.
It was just a few steps, but it's really cool to know how this automation can improve job productivity. We can, for example, collect data, and
after formatting and cleaning these data which are repetitive tasks, the tool records this process and applies it with the chosen command.